ASAR-CMS-Project/Canteen_Management_System/backend/core/serializers.py
from rest_framework import serializers
from .models import CanteenTiming

class CanteenTimingSerializer(serializers.ModelSerializer):
    class Meta:
        model = CanteenTiming
        fields = '__all__'